PYTHON-6019 Consolidate mongo_client module-level helpers into shared file
